Add series to post node.
Previous method for creating post was to run separate queries for post and series, because series had unilateral reference to post. This ended being very clunky and hard to extend. This change identifies series at the front end and adds series nodes to the referenced posts directly. This allows the post to run a single query for all relevant information. Same technique may be useful for pairing a uspecifier and its corresponding MD analysis file.